Refactored to have seperate text module
